explain how to add secrets
