+++ practice-app/practice_app/base_settings.py
@@ -11,9 +11,27 @@ https://docs.djangoproject.com/en/2.2/ref/settings/
 """
 
 import os
+import json
+
+from django.core.exceptions import ImproperlyConfigured
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+
+secrets = None
+
+
+def get_secret(name):
+    global secrets
+    if not secrets:
+        with open(os.path.join(BASE_DIR, 'secret.json')) as f:
+            secrets = json.loads(f.read())
+    try:
+        return secrets[name]
+    except KeyError:
+        raise ImproperlyConfigured(
+            "ImproperlyConfigured: Set {0} environment variable".format(name))
+
 
 # Application definition
